Report last breadcrumb for tests that fail with an uncaught exception

+++ test/test_run_test.js
@@ -322,8 +322,7 @@ describe('Test runner', function() {
       var process = runTest('suite_leave_breadcrumb', 'should leave breadcrumb');
       return when.promise(function(resolve) {
         process.on('message', function(message) {
-          if (message.type === 'breadcrumb') {
-            expect(message).property('message').to.be.equal('A breadcrumb');
+          if (message.type === 'breadcrumb' && message.message === 'A breadcrumb') {
             expect(message).property('trace').to.be.contain('suite_leave_breadcrumb.js:');
             process.kill();
             resolve();
